+++ abipc/abipc/IpcServer.h
@@ -38,12 +38,12 @@ class ServerConnection;
 class ServerMessageHandlers
 {
 private:
-    sa::CallableTable<size_t, void, ServerConnection&, const MessageBuffer&> handlers_;
+    sa::CallableTable<uint64_t, void, ServerConnection&, const MessageBuffer&> handlers_;
 public:
     template<typename _Msg>
     void Add(std::function<void(ServerConnection&, const _Msg& msg)>&& func)
     {
-        static constexpr size_t message_type = sa::StringHash(sa::TypeName<_Msg>::Get());
+        static constexpr uint64_t message_type = static_cast<uint64_t>(sa::StringHash(sa::TypeName<_Msg>::Get()));
         handlers_.Add(message_type, [handler = std::move(func)](ServerConnection& client, const MessageBuffer& buffer)
         {
             // Get() doesn't take const stuff, but the buffer isn't modified by it so just cast it away.
@@ -51,8 +51,8 @@ public:
             handler(client, packet);
         });
     }
-    bool Exists(size_t type) const { return handlers_.Exists(type); }
-    void Call(size_t type, ServerConnection& client, const MessageBuffer& buffer)
+    bool Exists(uint64_t type) const { return handlers_.Exists(type); }
+    void Call(uint64_t type, ServerConnection& client, const MessageBuffer& buffer)
     {
         handlers_.Call(type, client, buffer);
     }
@@ -84,7 +84,7 @@ public:
     template <typename _Msg>
     void Send(_Msg& msg)
     {
-        static constexpr size_t message_type = sa::StringHash(sa::TypeName<_Msg>::Get());
+        static constexpr uint64_t message_type = static_cast<uint64_t>(sa::StringHash(sa::TypeName<_Msg>::Get()));
         MessageBuffer buff;
         buff.type_ = message_type;
         Add(msg, buff);
@@ -95,7 +95,7 @@ public:
     template <typename _Msg>
     void SendTo(ServerConnection& client, _Msg& msg)
     {
-        static constexpr size_t message_type = sa::StringHash(sa::TypeName<_Msg>::Get());
+        static constexpr uint64_t message_type = static_cast<uint64_t>(sa::StringHash(sa::TypeName<_Msg>::Get()));
         MessageBuffer buff;
         buff.type_ = message_type;
         Add(msg, buff);
@@ -106,7 +106,7 @@ public:
     template <typename _Msg>
     void SendTo(uint32_t clientId, _Msg& msg)
     {
-        static constexpr size_t message_type = sa::StringHash(sa::TypeName<_Msg>::Get());
+        static constexpr uint64_t message_type = static_cast<uint64_t>(sa::StringHash(sa::TypeName<_Msg>::Get()));
         auto* client = GetConnection(clientId);
         if (client == nullptr)
             return;
